Task1/requirements.txt
matplotlib
mysql-connector-python
numpy
requests
tensorflow
keras